generative-manufacturing-server/pyproject.toml
[project]
name = "generative-manufacturing-server"
version = "0.1.0"
description = "MCP Server for Generative Manufacturing"
authors = [{name = "Technickel", email = "generative-manufacturing@technickel.dev"}]
requires-python = ">=3.10"
dependencies = [
    "mcp @ git+https://github.com/modelcontextprotocol/python-sdk@main",
    "qrcode[pil]>=8.0",
    "uvicorn>=0.34.0",
    "starlette>=0.46.0",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

